// src/app/auth/roles.enum.ts
// src/auth/roles.enum.ts
export enum Role {
  ADMIN = 'Admin',
  GURU = 'Guru',
  Murid = 'Murid',
  STAF = 'Staf',
  KEPALA_SEKOLAH = 'Kepala Sekolah',
  WALI_KELAS = 'Wali Kelas',
}

export enum StatusMapel {
  ONLINE = 'online',
  OFFLINE = 'offline',
}

export enum KelasEnum {
  X_RPL = 'X RPL',
  X_TKJ = 'X TKJ',
  XI_RPL = 'XI RPL',
  XI_TKJ = 'XI TKJ',
  XII_RPL = 'XII RPL',
  XII_TKJ = 'XII TKJ',
}